build: move test controller to skip build ignore
This dir has fixtures and we need to ignore them in build.
